Fix after self-review.
